feat(go/pkg/prometheus): replace ScrapeStream with ScrapeWithTransform
Add a per-sample transform hook for Prometheus-style metric relabeling and remove
the unused ScrapeStream callback API it supersedes.

- type SampleTransform func(Sample) (Sample, bool, error): keep (optionally mutated
  -- rewrite Name or mutate Labels in place), drop, or abort the scrape.
- ScrapeWithTransform(ctx, transform) (MetricFamilies, error) on the Prometheus
  interface; Scrape() is now ScrapeWithTransform(ctx, nil), so both share one
  assembly path.
- parseToMetricFamilies(text, transform): with a transform it sets ownLabels=true so
  the transform can mutate each sample's labels safely; with nil it keeps the
  no-allocation fast path. The transform runs before the assembler.
- Remove ScrapeStream + parseToStream (no production callers); stream.go held only
  ScrapeStream, so it is removed; the exposition-order/deferral caveat moves to
  ScrapeWithTransform's doc.
- Consolidate the parser tests into parse_test.go (stream_test.go no longer maps to a
  distinct surface); the parser benchmark measures the transform path.
