+++ src/go/pkg/prometheus/client.go
@@ -24,17 +24,18 @@ type (
 		// ScrapeSeries and parse prometheus format metrics
 		ScrapeSeries() (Series, error)
 		Scrape() (MetricFamilies, error)
-		// ScrapeStream scrapes and invokes onSample for every sample — a flat
-		// [Sample] stream before typed-family assembly — with the selector (if
-		// any) already applied; onHelp, if non-nil, receives per-family HELP. It
-		// stops and returns the first error from a callback.
+		// ScrapeWithTransform scrapes, runs transform on every sample (a flat
+		// [Sample] stream before typed-family assembly, with the selector already
+		// applied), then assembles the kept, transformed samples into MetricFamilies.
+		// A nil transform behaves exactly like Scrape. The result aliases reused
+		// buffers, valid until the next scrape on this instance (same as Scrape).
 		//
-		// Order is exposition order, with one exception: a _sum/_count sample
-		// whose family type is not yet known (its # TYPE, first bucket, or first
-		// quantile has not appeared) is deferred and emitted once the type
-		// resolves, or at end of stream — so it may arrive after a later,
-		// unrelated sample. Each sample's Kind and FamilyType are always correct.
-		ScrapeStream(ctx context.Context, onHelp func(name, help string), onSample func(Sample) error) error
+		// Samples reach transform in exposition order, with one exception: a
+		// _sum/_count sample whose family type is not yet known (its # TYPE, first
+		// bucket, or first quantile has not appeared) is deferred and delivered once
+		// the type resolves, or at end of scrape — so it may arrive after a later,
+		// unrelated sample. A per-sample (stateless) transform is unaffected.
+		ScrapeWithTransform(ctx context.Context, transform SampleTransform) (MetricFamilies, error)
 		HTTPClient() *http.Client
 	}
 
@@ -95,13 +96,17 @@ func (p *prometheus) ScrapeSeries() (Series, error) {
 }
 
 func (p *prometheus) Scrape() (MetricFamilies, error) {
+	return p.ScrapeWithTransform(context.Background(), nil)
+}
+
+func (p *prometheus) ScrapeWithTransform(ctx context.Context, transform SampleTransform) (MetricFamilies, error) {
 	p.buf.Reset()
 
-	if err := p.fetch(context.Background(), p.buf); err != nil {
+	if err := p.fetch(ctx, p.buf); err != nil {
 		return nil, err
 	}
 
-	return p.parser.parseToMetricFamilies(p.buf.Bytes())
+	return p.parser.parseToMetricFamilies(p.buf.Bytes(), transform)
 }
 
 func (p *prometheus) fetch(ctx context.Context, w io.Writer) error {
